chore(TiktoAds): move sdk loader to separate file
